+++ CountlyConnectionManager.m
@@ -419,7 +419,7 @@ static dispatch_once_t onceToken;
             }
             else
             {
-                CLY_LOG_D(@"%s, request:[ <%p> ] failed! response:[ %@ ]", __FUNCTION__, request, [data cly_stringUTF8]);
+                CLY_LOG_E(@"%s, request:[ <%p> ] failed! response:[ %@ ]", __FUNCTION__, request, [data cly_stringUTF8]);
 
                 self.hasAnyRequestFailed = YES; // Mark that a request has failed
 
@@ -434,6 +434,7 @@ static dispatch_once_t onceToken;
                 }
 
                 [CountlyHealthTracker.sharedInstance logFailedNetworkRequestWithStatusCode:((NSHTTPURLResponse*)response).statusCode errorResponse: [data cly_stringUTF8]];
+                [CountlyHealthTracker.sharedInstance recordLogCode:@"e301"];
                 [CountlyHealthTracker.sharedInstance saveState];
                 self.startTime = nil;
                 atomic_store(&self->_isProcessingQueue, NO);