+++ CountlyViewTrackingInternal.m
@@ -156,9 +156,11 @@ NSString* const kCountlyVTKeyDur      = @"dur";
 - (NSString *)startView:(NSString *)viewName segmentation:(NSDictionary *)segmentation
 {
     CLY_LOG_I(@"%s %@ %@", __FUNCTION__, viewName, segmentation);
+    [CountlyHealthTracker.sharedInstance recordUsage:@"views" method:@"start"];
 #if (TARGET_OS_IOS || TARGET_OS_TV)
     if (self.isAutoViewTrackingActive) {
         CLY_LOG_W(@"%s Manually start view tracking is not allowed when automatic tracking is enabled!", __FUNCTION__);
+        [CountlyHealthTracker.sharedInstance recordLogCode:@"w740"];
         return nil;
     }
 #endif
@@ -169,9 +171,11 @@ NSString* const kCountlyVTKeyDur      = @"dur";
 - (NSString *)startAutoStoppedView:(NSString *)viewName segmentation:(NSDictionary *)segmentation
 {
     CLY_LOG_I(@"%s %@ %@", __FUNCTION__, viewName, segmentation);
+    [CountlyHealthTracker.sharedInstance recordUsage:@"views" method:@"startAuto"];
 #if (TARGET_OS_IOS || TARGET_OS_TV)
     if (self.isAutoViewTrackingActive) {
         CLY_LOG_W(@"%s Manually start view tracking is not allowed when automatic tracking is enabled!", __FUNCTION__);
+        [CountlyHealthTracker.sharedInstance recordLogCode:@"w740"];
         return nil;
     }
 #endif
@@ -179,12 +183,30 @@ NSString* const kCountlyVTKeyDur      = @"dur";
     return viewID;
 }
 
+// Backing call for the deprecated `recordView:` API. Mirrors `startAutoStoppedView:` but records
+// the usage under `start:d` so deprecated view calls are distinguishable from the current API.
+- (NSString *)recordViewDeprecated:(NSString *)viewName segmentation:(NSDictionary *)segmentation
+{
+    CLY_LOG_I(@"%s %@ %@", __FUNCTION__, viewName, segmentation);
+    [CountlyHealthTracker.sharedInstance recordUsage:@"views" method:@"start:d"];
+#if (TARGET_OS_IOS || TARGET_OS_TV)
+    if (self.isAutoViewTrackingActive) {
+        CLY_LOG_W(@"%s Manually start view tracking is not allowed when automatic tracking is enabled!", __FUNCTION__);
+        [CountlyHealthTracker.sharedInstance recordLogCode:@"w740"];
+        return nil;
+    }
+#endif
+    return [self startViewInternal:viewName customSegmentation:segmentation isAutoStoppedView:YES];
+}
+
 - (void)stopViewWithName:(NSString *)viewName segmentation:(NSDictionary *)segmentation
 {
     CLY_LOG_I(@"%s %@ %@", __FUNCTION__, viewName, segmentation);
+    [CountlyHealthTracker.sharedInstance recordUsage:@"views" method:@"stop"];
 #if (TARGET_OS_IOS || TARGET_OS_TV)
     if (self.isAutoViewTrackingActive) {
         CLY_LOG_W(@"%s Manually stop view tracking is not allowed when automatic tracking is enabled!", __FUNCTION__);
+        [CountlyHealthTracker.sharedInstance recordLogCode:@"w740"];
         return;
     }
 #endif
@@ -195,9 +217,11 @@ NSString* const kCountlyVTKeyDur      = @"dur";
 - (void)stopViewWithID:(NSString *)viewID segmentation:(NSDictionary *)segmentation
 {
     CLY_LOG_I(@"%s %@ %@", __FUNCTION__, viewID, segmentation);
+    [CountlyHealthTracker.sharedInstance recordUsage:@"views" method:@"stop"];
 #if (TARGET_OS_IOS || TARGET_OS_TV)
     if (self.isAutoViewTrackingActive) {
         CLY_LOG_W(@"%s Manually stop view tracking is not allowed when automatic tracking is enabled!", __FUNCTION__);
+        [CountlyHealthTracker.sharedInstance recordLogCode:@"w740"];
         return;
     }
 #endif
@@ -210,6 +234,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
 #if (TARGET_OS_IOS || TARGET_OS_TV)
     if (self.isAutoViewTrackingActive) {
         CLY_LOG_W(@"%s Manually pause view tracking is not allowed when automatic tracking is enabled!", __FUNCTION__);
+        [CountlyHealthTracker.sharedInstance recordLogCode:@"w740"];
         return;
     }
 #endif
@@ -222,6 +247,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
 #if (TARGET_OS_IOS || TARGET_OS_TV)
     if (self.isAutoViewTrackingActive) {
         CLY_LOG_W(@"%s Manually resume view tracking is not allowed when automatic tracking is enabled!", __FUNCTION__);
+        [CountlyHealthTracker.sharedInstance recordLogCode:@"w740"];
         return;
     }
 #endif
@@ -234,6 +260,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
 #if (TARGET_OS_IOS || TARGET_OS_TV)
     if (self.isAutoViewTrackingActive) {
         CLY_LOG_W(@"%s Manually stop view tracking is not allowed when automatic tracking is enabled!", __FUNCTION__);
+        [CountlyHealthTracker.sharedInstance recordLogCode:@"w740"];
         return;
     }
 #endif
@@ -261,6 +288,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
     
     UIViewController* topVC = CountlyCommon.sharedInstance.topViewController;
     NSString* viewTitle = [self titleForViewController:topVC];
+    [CountlyHealthTracker.sharedInstance recordUsage:@"views" method:@"start:a"];
     [self startViewInternal:viewTitle customSegmentation:nil];
 }
 
@@ -716,7 +744,10 @@ NSString* const kCountlyVTKeyDur      = @"dur";
     }
     
     if (!isException)
+    {
+        [CountlyHealthTracker.sharedInstance recordUsage:@"views" method:@"start:a"];
         [self startViewInternal:viewTitle customSegmentation:nil];
+    }
 }